requirements.txt
Django>=2.0,<3.0
scikit-learn
mord
djangorestframework==3.11.0
numpy
pandas
nltk
